python-worker/requirements.txt
opencv-python==4.11.0.86
psycopg-binary==3.2.9
python-dotenv==1.0.0
APScheduler==3.10.4
requests==2.31.0
pillow==11.2.1

# Testing dependencies
pytest==8.4.0
pytest-mock==3.12.0
pytest-asyncio==1.0.0
pytest-cov==6.1.1
